[elasticsearch]: master node performs data role during bootstrap
There was a race condition.
During bootstrapping a cluser, elastic wants to create some system
indexes. Before it does so, the cluster is green and the Terraform
module proceeded with creating data nodes. But if the cluster became
red, Terraform didn't proceed with the data nodes and failed.

Now when a sigle master node bootstraps a cluster, it also is a data
node. The cluster will be yellow, but that's enough for a healthcheck.